Move UserScoreable trait
