+++ app/Models/User.php
@@ -174,7 +174,7 @@ use Request;
 class User extends Model implements AfterCommit, AuthenticatableContract, HasLocalePreference, Indexable
 {
     use Elasticsearch\UserTrait, Traits\UserStore;
-    use Authenticatable, HasApiTokens, Memoizes, Traits\Reportable, UserAvatar, UserScoreable, Validatable;
+    use Authenticatable, HasApiTokens, Memoizes, Traits\Reportable, UserAvatar, Traits\UserScoreable, Validatable;
 
     protected $table = 'phpbb_users';
     protected $primaryKey = 'user_id';
